% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_plots.R
\name{get_odds}
\alias{get_odds}
\title{get the df odds}
\usage{
get_odds(va.list, df)
}
\arguments{
\item{va.list}{the sep points list}

\item{df}{the dataframe need to transfer}
}
\description{
this function will get the df odds
}
\examples{
get_odds(va.list,df)
}

