% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_plots.R
\name{get_sep_point}
\alias{get_sep_point}
\title{get sep point}
\usage{
get_sep_point(vec, minnums, sepnums)
}
\arguments{
\item{vec}{the numeric vector}

\item{minnums}{the minimul observe numbers}

\item{sepnums}{the sep points}
}
\description{
this function will get sep point
}
\examples{
single_value_count(p2p,"def")
}

