% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_plots.R
\name{numvar_animation}
\alias{numvar_animation}
\title{get a animate numeric plots}
\usage{
numvar_animation(value, target, timestp, varname,
  dir = "C:/Users/uij/Documents/R/image/", nums = 500, nsp = 20,
  tnums = 5000, npics = 300, lg = F, miss = -9999)
}
\arguments{
\item{value}{the dataframe}

\item{target}{the label variable}

\item{timestp}{the label variable}

\item{varname}{the label variable}

\item{dir}{the label variable}
}
\description{
this function will get a animate numeric plots
}
\examples{
numvar_animation(value,target,timestp,varname,dir,nums=500,nsp=20,tnums=5000,npics=300,lg=F)
}

