% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single variable analysis.R
\name{single_catvar_analysis}
\alias{single_catvar_analysis}
\title{doing category variable analysis}
\usage{
single_catvar_analysis(value, target, timestp, varname, ord = T, tcp = NULL,
  miss = -9999)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}
}
\description{
this function will doing category variable analysis
}
\examples{
single_catvar_analysis(value,target,timestp,varname,ord=T,tcp=NULL)
}

