% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single variable analysis.R
\name{single_numvar_analysis}
\alias{single_numvar_analysis}
\title{doing single numeric variable analysis}
\usage{
single_numvar_analysis(value, target, timestp, varname, tcp = NULL,
  nums = 500, nsp = 20, lg = F, ord = F, miss = -9999)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}
}
\description{
this function will doing single numeric variable analysis
}
\examples{
single_numvar_analysis(value,target,timestp,varname,tcp=NULL,nums=500,nsp=20,lg=F,ord=F)
}

