% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single variable analysis.R
\name{time_trend_analysis}
\alias{time_trend_analysis}
\title{doing time trends analysis}
\usage{
time_trend_analysis(target, timestp, nums = 500, nsp = 20, events = NULL,
  timepoint = NULL, miss = -9999)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}
}
\description{
this function will doing time trends analysis
}
\examples{
time_trend_analysis(target,timestp,nums=500,nsp=20,events=NULL,timepoint=NULL)
}

