% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation_plots.R
\name{varIV_animation}
\alias{varIV_animation}
\title{get a animate catgory variable plots}
\usage{
varIV_animation(adf, varn.change, var.clu, dir, lowest = 500, tnums = 10000,
  npics = 300, miss = -9999)
}
\arguments{
\item{dir}{the label variable}

\item{value}{the dataframe}

\item{target}{the label variable}

\item{timestp}{the label variable}

\item{varname}{the label variable}
}
\description{
this function will get a animate catgory variable plots
}
\examples{
catvar_animation(value,target,timestp,varname,dir,tnums=5000,npics=300)
}

